made routes more modular with Blueprint. seperated database connection and models into separate folders to work on
