+++ backend/database/__init__.py
@@ -1,0 +1,1 @@
+from .connection import databaseConnection
