+++ backend/routes/__init__.py
@@ -1,0 +1,2 @@
+from .health import health_bp
+from .sentiment import sentiment_bp
